Extract import parsing from ProgrammingExercise
